+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.10.0] - 2026-04-14
+
 ### Added
 
 - Icon for Jupyter Notebook
@@ -293,7 +295,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release
 
-[Unreleased]: https://github.com/pandier/intellij-discord-rp/compare/v1.9.1...HEAD
+[Unreleased]: https://github.com/pandier/intellij-discord-rp/compare/v1.10.0...HEAD
+[1.10.0]: https://github.com/pandier/intellij-discord-rp/compare/v1.9.1...v1.10.0
 [1.9.1]: https://github.com/pandier/intellij-discord-rp/compare/v1.9.0...v1.9.1
 [1.9.0]: https://github.com/pandier/intellij-discord-rp/compare/v1.8.0...v1.9.0
 [1.8.0]: https://github.com/pandier/intellij-discord-rp/compare/v1.7.4...v1.8.0